pyproject.toml
[project]
name = "mcp-adapters-explore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain-gigachat>=0.3.6",
    "langchain-mcp-adapters>=0.0.3",
    "langchain-openai>=0.3.7",
    "langgraph>=0.3.5",
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.1",
]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"